remove codable data helpers
